Populate with Rust and fidl
This commit populates the repo with Rust source (a thin wrapper around
xi-core-lib that provides communication through fidl) and the fidl
interfaces. The start of the front end will be in a followup CL.

The Cargo.toml for the Rust binary assumes that xi-core-lib and
dependent crates are published to crates.io.

Change-Id: I764955b0c81e53b29e94c9ee1610f59eacdb3ddd